+++ mis_builder/tests/__init__.py
@@ -17,3 +17,4 @@ from . import test_subreport
 from . import test_target_move
 from . import test_utc_midnight
 from . import test_mis_report_instance_annotation
+from . import test_detail_groupby